# chore: parallelize mprocs lightning channel opening