--- a/devimint/Cargo.toml
+++ b/devimint/Cargo.toml
@@ -49,6 +49,7 @@ fedimintd = { path = "../fedimintd" }
 fs-lock = "0.1.3"
 futures = { workspace = true }
 hex = { workspace = true }
+itertools = { workspace = true }
 ln-gateway = { package = "fedimint-ln-gateway", path = "../gateway/ln-gateway" }
 nix = { version = "0.29.0", features = ["signal"] }
 rand = { workspace = true }