On my local machine (Apple M1 Pro, 32GB RAM) the mprocs startup times before/after are:

Before (master branch):
Attempt 1: 93.3s
Attempt 2: 95.5s
Attempt 3: 94.0s
Attempt 4: 93.8s
Attempt 5: 93.5s
Average: 94.0s

After (PR branch):
Attempt 1: 65.2s
Attempt 2: 64.0s
Attempt 3: 63.7s
Attempt 4: 65.7s
Attempt 5: 64.3s
Average: 64.6s

So this is roughly 30% faster and helps to undo some of the damage caused by merging the LDK gateway