@dpc 2024-08-10
https://github.com/fedimint/fedimint/pull/5826 instead?