@dpc 2024-08-10
Well, https://github.com/fedimint/fedimint/pull/5826 doesn't work, so whatevs. ;D