@Kodylow 2024-08-12
@dpc eric said on the call you might have a concern here, can you post if so otherwise can you land?