@dpc 2024-08-19
@tvolk131 

> But without this change, we're running the test (f()) before waiting for the federation to be finalized. Isn't that a race condition?

The idea was that particular tests should wait for things it actually needs, so that we don't have to pessimize in ever test.